New designs and changed many things
The anime card's design has been changed for a new one, showing more information and with a better design (According to the author's taste),  the same happened with the profile selector, this with the intention of having more consistence in the whole app, then the other card's will be changed to the new style. The font has been changed to calibri,. The addToList button has suffered some changes removing useless code, improving the design and adding some animations to give more feedback to the user.  The function used to delete and set cookies (deleteCookie and setCookie) have been replaced to the native implementation and now it's used just the getCookie functions, this with the intention to have a better control of the cookies because there are times that the cookies are not deleted and doing this, i can change it at time until fix it. And other small fixes and adding